+++ src/igor/igor/renderer/iRenderer.cpp
@@ -554,9 +554,8 @@ namespace igor
         _data->_lastRenderDataSetUsed = iRenderDataSet::NoDataSet;
         _data->_currentShader.reset();
 
-        ////////////// generate textures //////////
+        ////////////// generate fallback texture //////////////
         iParameters paramFallback({{IGOR_RESOURCE_PARAM_ID, iaUUID()},
-                                   {IGOR_RESOURCE_PARAM_ALIAS, "igor_texture_fallback"},
                                    {IGOR_RESOURCE_PARAM_TYPE, IGOR_RESOURCE_TEXTURE},
                                    {IGOR_RESOURCE_PARAM_CACHE_MODE, iResourceCacheMode::Keep},
                                    {IGOR_RESOURCE_PARAM_GENERATE, true},
@@ -567,35 +566,6 @@ namespace igor
                                    {IGOR_RESOURCE_PARAM_TEXTURE_HEIGHT, 128}});
 
         _data->_fallbackTexture = iResourceManager::getInstance().loadResource<iTexture>(paramFallback);
-        iResourceManager::getInstance().addToDictionary("", "igor_texture_fallback", _data->_fallbackTexture->getID());
-
-        createSolidColorTexture("igor_texture_white", iaColor4f::white);
-        createSolidColorTexture("igor_texture_black", iaColor4f::black);
-        createSolidColorTexture("igor_texture_red", iaColor4f::red);
-        createSolidColorTexture("igor_texture_blue", iaColor4f::blue);
-        createSolidColorTexture("igor_texture_green", iaColor4f::green);
-        createSolidColorTexture("igor_texture_cyan", iaColor4f::cyan);
-        createSolidColorTexture("igor_texture_magenta", iaColor4f::magenta);
-        createSolidColorTexture("igor_texture_yellow", iaColor4f::yellow);
-        createSolidColorTexture("igor_texture_transparent", iaColor4f::transparent);
-    }
-
-    iResourcePtr iRenderer::createSolidColorTexture(const iaString &alias, const iaColor4f &color)
-    {
-        iParameters paramWhite({{IGOR_RESOURCE_PARAM_ID, iaUUID()},
-                                {IGOR_RESOURCE_PARAM_ALIAS, alias},
-                                {IGOR_RESOURCE_PARAM_TYPE, IGOR_RESOURCE_TEXTURE},
-                                {IGOR_RESOURCE_PARAM_CACHE_MODE, iResourceCacheMode::Keep},
-                                {IGOR_RESOURCE_PARAM_GENERATE, true},
-                                {IGOR_RESOURCE_PARAM_TEXTURE_PATTERN, iTexturePattern::SolidColor},
-                                {IGOR_RESOURCE_PARAM_PRIMARY_COLOR, color},
-                                {IGOR_RESOURCE_PARAM_TEXTURE_WIDTH, 1},
-                                {IGOR_RESOURCE_PARAM_TEXTURE_HEIGHT, 1}});
-
-        auto resource = iResourceManager::getInstance().loadResource(paramWhite);
-        iResourceManager::getInstance().addToDictionary("", alias, resource->getID());
-
-        return resource;
     }
 
     void iRenderer::deinit()