+++ src/igor/igor/renderer/iRenderer.h
@@ -975,13 +975,6 @@ namespace igor
         \param color the color to draw with
         */
         void drawBoxInternal(const iAABoxf &box, const iaColor4f &color);
-
-        /*! creates solid color 1x1 texture and registers it with the dictionary
-
-        \param alias the alias of the texture
-        \param color the color of the texture
-        */
-        iResourcePtr createSolidColorTexture(const iaString& alias, const iaColor4f& color);
     };
 
 #include <igor/renderer/iRenderer.inl>