+++ src/docs/changelog.md
@@ -12,6 +12,8 @@ Changelog                           {#changelog}
 - #375 cleaned up iPixmap interface and reduced it to a minimum
 - #362 removed iwidget::getKeyboardFocusWidget. It was not used
 - #306 using ChaiScript for scripting
+- added offset to camera component
+- #457 TODO
 
 0.45.0
 ------